Place raven error handler in a separate file
